Move memory-backed KeyValue implementation into unit tests where it belongs; enabled by elimination of templates.
