+++ include/KeyValue.h
@@ -75,125 +75,6 @@ public:
     Status put(CollectionHandle coll, const std::string& key, const std::string& value);
 };
 
-// Trivial in-memory KeyValue::DB implementation used in unit tests.
-namespace Mem {
-    class Iterator : public KeyValue::Iterator {
-        std::map<std::string,std::string> data_;
-        std::map<std::string,std::string>::const_iterator it_;
-        friend class Reader;
-
-    public:
-        Status next(std::string& key, std::string& value) override {
-            if (it_ == data_.end()) return Status::NotFound();
-            key = it_->first;
-            value = it_->second;
-            it_++;
-            return Status::OK();
-        }
-    };
-
-    class Reader : public KeyValue::Reader {
-        std::vector<std::map<std::string,std::string>> data_;
-        friend class DB;
-
-    public:
-        Status get(CollectionHandle _coll, const std::string& key, std::string& value) const override {
-            auto coll = reinterpret_cast<uint64_t>(_coll);
-            assert(coll < data_.size());
-            const auto& m = data_[coll];
-            auto p = m.find(key);
-            if (p == m.end()) return Status::NotFound("key", key);
-            value = p->second;
-            return Status::OK();
-        }
-
-        Status iterator(CollectionHandle _coll, std::unique_ptr<KeyValue::Iterator>& it) const override {
-            auto coll = reinterpret_cast<uint64_t>(_coll);
-            assert(coll < data_.size());
-            auto it2 = std::make_unique<Iterator>();
-            it2->data_ = data_[coll];
-            it2->it_ = it2->data_.begin();
-            it.reset(it2.release());
-            return Status::OK();
-        }
-
-        Status iterator(CollectionHandle _coll, const std::string& key, std::unique_ptr<KeyValue::Iterator>& it) const override {
-            auto coll = reinterpret_cast<uint64_t>(_coll);
-            assert(coll < data_.size());
-            auto it2 = std::make_unique<Iterator>();
-            it2->data_ = data_[coll];
-            it2->it_ = it2->data_.lower_bound(key);
-            it.reset(it2.release());
-            return Status::OK();
-        }
-    };
-
-    class DB;
-
-    class WriteBatch : public KeyValue::WriteBatch {
-        std::vector<std::map<std::string,std::string>> data_;
-        DB* db_;
-        friend class DB;
-
-    public:
-        Status put(CollectionHandle _coll, const std::string& key, const std::string& value) override;
-        Status commit() override;
-    };
-
-    class DB : public KeyValue::DB {
-        std::map<std::string,uint64_t> collections_;
-        std::vector<std::map<std::string,std::string>> data_;
-        friend class WriteBatch;
-
-    public:
-        DB(const std::vector<std::string>& collections) {
-            for (uint64_t i = 0; i < collections.size(); i++) {
-                assert(collections_.find(collections[i]) == collections_.end());
-                collections_[collections[i]] = i;
-            }
-            data_ = std::vector<std::map<std::string,std::string>>(collections_.size());
-        }
-
-        Status collection(const std::string& name, CollectionHandle& coll) const override {
-            auto p = collections_.find(name);
-            if (p != collections_.end()) {
-                coll = reinterpret_cast<CollectionHandle>(p->second);
-                return Status::OK();
-            }
-            return Status::NotFound("KeyValue::Mem::collection", name);
-        }
-
-        Status create_collection(const std::string& name) override {
-            if (collections_.find(name) != collections_.end()) {
-                return Status::Exists("key-value collection already exists", name);
-            }
-            collections_[name] = data_.size();
-            data_.emplace_back();
-            return Status::OK();
-        }
-
-        Status current(std::unique_ptr<KeyValue::Reader>& reader) const override {
-            auto p = std::make_unique<KeyValue::Mem::Reader>();
-            p->data_ = data_;
-            reader = std::move(p);
-            return Status::OK();
-        }
-
-        Status begin_writes(std::unique_ptr<KeyValue::WriteBatch>& writes) override {
-            auto p = std::make_unique<Mem::WriteBatch>();
-            p->db_ = this;
-            p->data_ = std::vector<std::map<std::string,std::string>>(data_.size());
-            writes = std::move(p);
-            return Status::OK();
-        }
-
-        void wipe() {
-            collections_.clear();
-            data_.clear();
-        }
-    };
-}
-
 }}
 
 #endif
